backend/requirements.txt
python-dotenv
pytest
yt-dlp
demucs
soundfile
librosa
numpy
scipy
pyloudnorm
fastapi
uvicorn
python-multipart
google-cloud-firestore
google-cloud-storage
